test: break court tests into more specific files
This makes the domain of each set of tests clearer.
